requirements.txt
alembic==1.5.8
blinker==1.4
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==4.1.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
gevent==21.1.2
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.7
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.3.0
